swarm/network/stream: addressed PR comments @janos
